fix: normalize messages in all engine paths before chat template
Extract _normalize_messages() to shared message_utils.py module.
Add calls in BatchedEngine.chat() and BatchedEngine.stream_chat().

Qwen 3.5 templates reject developer role, consecutive same-role
messages, and system messages out of position [0].
